sheetBS

+++ Sheet/Arrays/Binary Search/median2sortedarrays.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 class Solution{
 public:
-    int medianofsortedarrays(vector<int>nums1,vector<int>nums2){
+    double medianofsortedarrays(vector<int>nums1,vector<int>nums2){
         if(nums2.size()<nums1.size()) return medianofsortedarrays(nums2,nums1);  // nums1 as search space smaller n1. 
         int n1=nums1.size();  // nums1 should be smaller and cut1
         int n2=nums2.size();
